requirements.txt
chromadb==1.0.20
google-genai==1.35.0
sentence-transformers==5.1.0
python-dotenv==1.1.1